requirements.txt
scikit-learn==1.4.0
pickle-mixin==1.0.2
PyYAML==6.0.1
fastapi==0.109.1
uvicorn==0.27.0.post1
pandas==2.2.0
numpy==1.26.3
pydantic==2.6.0
pydantic_core==2.16.1

